build: move sched-simple rlist code to common/librlist
Problem: the "rlist" code in sched-simple may be able to serve as
an interim Rv1 parsing and manipulation library, but the code must
first be moved where multiple modules and/or utilities can link
against it.

Move the rlist code to src/common/librlist and link sched-simple with
the new librlist convienience library.
