+++ src/common/Makefile.am
@@ -21,7 +21,8 @@ SUBDIRS = \
 	libdebugged \
 	libterminus \
 	libcontent \
-	libhostlist
+	libhostlist \
+	librlist
 
 AM_CFLAGS = $(WARNING_CFLAGS) $(CODE_COVERAGE_CFLAGS)
 AM_LDFLAGS = $(CODE_COVERAGE_LIBS)
